+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to Semantic Versioning (https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Default DFM to `standard` and consolidate the bundled JLCPCB PDK.
+
 ## [0.4.48] - 2026-09-04
 
 ### Added